+++ hurricane/src/clock.c
@@ -28,10 +28,10 @@ void hc_clock_new(hc_clock *clock) {
     //gettimeofday(&clock->last_time, NULL);
 }
 
-float hc_clock_step(hc_clock *clock) {
+double hc_clock_step(hc_clock *clock) {
     //struct timeval now;
     //gettimeofday(&now, NULL);
-    //float delta = (now.tv_sec - clock->last_time.tv_sec) + (now.tv_usec - clock->last_time.tv_usec) / 1000000.0;
+    //double delta = (now.tv_sec - clock->last_time.tv_sec) + (now.tv_usec - clock->last_time.tv_usec) / 1000000.0;
     //clock->last_time = now;
     return 0.1;
 }
