+++ hurricane/src/loader/obj.c
@@ -22,7 +22,7 @@ void hc_geometry_from_obj(char *file, hc_geometry *output) {
   int face_count = 0;
   int face_pointer = 0;
   
-  output->vertices = malloc(sizeof(float) * vertex_buffer_size * 3);
+  output->vertices = malloc(sizeof(double) * vertex_buffer_size * 3);
   output->faces = malloc(sizeof(int) * face_buffer_size * 3);
   output->colors = malloc(sizeof(int) * 10000 * 3); //todo: needs fix?
 
@@ -34,11 +34,11 @@ void hc_geometry_from_obj(char *file, hc_geometry *output) {
 
     if (vertex_buffer_size <= vertex_count) {
       vertex_buffer_size <<= 2;
-      output->vertices = realloc(output->vertices, sizeof(float) * vertex_buffer_size * 3);
+      output->vertices = realloc(output->vertices, sizeof(double) * vertex_buffer_size * 3);
     }
     if (face_buffer_size <= face_count) {
       face_buffer_size <<= 2;
-      output->faces = realloc(output->faces, sizeof(float) * face_buffer_size * 3);
+      output->faces = realloc(output->faces, sizeof(double) * face_buffer_size * 3);
     }
 
     if (line[0] == 'v') {
